# app/static/makeascene.openapi.yaml
openapi: 3.1.0
info:
  title: makeascene
  description: |
    The API used for the frontend and third-party for Make A Scene
  version: 1.3.0
...
paths:
  /users/me/tags/{id}:
    delete:
      summary: Unsubscribe tag
      deprecated: false
      description: Remove a tag if it exists.
      operationId: deleteTag
      tags:
        - Users
        - Tags
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security:
        - BearerAuth: []
    post:
      summary: Subscribe tag
      deprecated: false
      description: Saves a tag if it exists.
      operationId: saveTag
      tags:
        - Users
        - Tags
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security:
        - BearerAuth: []
  /users/me/tags:
    get:
      summary: Get all subscribed tags
      deprecated: false
      description: Get the tags you saved.
      operationId: getTags
      tags:
        - Users
        - Tags
      parameters:
        - name: page
          in: query
          description: ''
          required: false
          example: 1
          schema:
            type: integer
        - name: page_size
          in: query
          description: ''
          required: false
          example: 25
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagQueryResponse'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security:
        - BearerAuth: []
  /users/me/tags/block/{id}:
    post:
      summary: Block tag
      deprecated: false
      description: Blocks a tag if it exists.
      operationId: blockTag
      tags:
        - Users
        - Tags
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security:
        - BearerAuth: []
    delete:
      summary: Unblock Tag
      deprecated: false
      description: Unblocks a tag if it exists and is blocked.
      operationId: unblockTag
      tags:
        - Users
        - Tags
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security:
        - BearerAuth: []
  /users/me/tags/block:
    get:
      summary: Get all blocked tags
      deprecated: false
      description: Returns all the tags a user blocked
      operationId: getBlockedTags
      tags:
        - Users
        - Tags
      parameters:
        - name: page
          in: query
          description: ''
          required: false
          example: 1
          schema:
            type: integer
        - name: page_size
          in: query
          description: ''
          required: false
          example: 25
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
          headers: {}
      security:
        - BearerAuth: []
  /users/me:
    get:
      summary: Get info about yourself
      deprecated: false
      description: Retrieve info about yourself
      operationId: getInfo
      tags:
        - Users
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security:
        - BearerAuth: []
    patch:
      summary: Change info about yourself
      deprecated: false
      description: Change info about yourself
      operationId: changeInfo
      tags:
        - Users
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchUserRequest'
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security:
        - BearerAuth: []
    delete:
      summary: Delete yourself
      deprecated: false
      description: Deletes your account
      operationId: deleteUser
      tags:
        - Users
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteUserResponse'
          headers: {}
        '401':
          $ref: '#/components/responses/Unauthorized'
          description: Authentication failed
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security:
        - BearerAuth: []
...
  /users:
    post:
      summary: Create user
      deprecated: false
      description: >-
        Registers a new user with email and password. Fails if the email is
        already taken or if input validation fails.
      operationId: createUser
      tags:
        - Users
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            examples:
              ValidRequest:
                value:
                  email: max.mustermann@example.com
                  password: securePassword123
                summary: Valid registration
              WeakPassword:
                value:
                  email: max.mustermann@example.com
                  password: short
                summary: Password too short (triggers 422)
              InvalidEmail:
                value:
                  email: no-email
                  password: securePassword123
                summary: Invalid email format (triggers 422)
        required: true
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: User created successfully.
          headers: {}
        '409':
          description: A user with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: conflict
                message: User already exists.
          headers: {}
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
          description: Validation error – one or more fields failed validation
        '500':
          $ref: '#/components/responses/InternalServerError'
          description: ''
      security: []
...
components:
  schemas:
...
    UserInfoResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID
        email:
          type: string
          format: email
          examples:
            - test@test.com
        mature:
          type: boolean
          default: true
        email_preference:
          type: boolean
          default: true
      required:
        - mature
        - email_preference
        - email
        - id
    PatchUserRequest:
      type: object
      properties:
        mature:
          type: boolean
          default: true
        email_preference:
          type: boolean
          default: true
    CreateUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          examples:
            - max.mustermann@example.com
        password:
          type: string
          format: password
          minLength: 8
          examples:
            - securePassword123
    DeleteUserResponse:
      type: object
      properties:
        message:
          type: string
          examples:
            - Deletion was successful.
      required:
        - message
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          examples:
            - Request body validation failed
        message:
          type: string
          examples:
            - A human-readable description of the error.
    TagQueryResponse:
      type: object
      properties:
        tags:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                description: ID
              name:
                type: string
                description: name
                examples:
                  - test tag
              parent:
                type:
                  - integer
                  - 'null'
                description: The parent tag
            required:
              - id
              - name
          description: tag
      required:
        - tags
    TagResponse:
      type: object
      properties:
        message:
          type: string
          examples:
            - Tag saved.
      required:
        - message
...
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                examples:
                  - unauthorized
              message:
                type: string
                examples:
                  - Invalid credentials.
            required:
              - error
              - message
          examples:
            Success:
              summary: Success
              value:
                error: unauthorized
                message: Invalid credentials.
    UnprocessableEntity:
      description: Validation error – one or more fields failed validation
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                examples:
                  - Request body validation failed
              fields:
                type: array
                items:
                  type: string
                examples:
                  - - '''password'' is too short (minimum length: 8)'
                    - '''email'' is not a valid email address'
          examples:
            Example 1:
              summary: Example 1
              value:
                error: Request body validation failed
                fields:
                  - '''password'' is too short (minimum length: 8)'
                  - '''email'' is not a valid email address'
    InternalServerError:
      description: ''
      content:
        application/octet-stream:
          schema:
            type: object
            properties: {}
          examples:
            Example 1:
              summary: Example 1
              value: {}
    NotFound:
      description: ''
      content:
        application/json:
          schema:
            title: ''
            type: object
            properties:
              error:
                type: string
                examples:
                  - Not found
              message:
                type: string
                examples:
                  - Tag was not found.
            required:
              - error
              - message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
servers:
  - url: http://127.0.0.1:5000/v1
    description: http://127.0.0.1:5000/v1
security:
  - BearerAuth: []